refactor: 댓글 생성할 때 내용 확인하는 역할을 Comment 객체에서 하도록 변경
