+++ src/main/java/com/hotnerds/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.hotnerds.comment.repository;
 
-import com.hotnerds.post.domain.comment.Comment;
+import com.hotnerds.comment.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
