+++ src/main/java/com/hotnerds/post/domain/Post.java
@@ -3,8 +3,8 @@ package com.hotnerds.post.domain;
 import com.hotnerds.common.BaseTimeEntity;
 import com.hotnerds.post.domain.like.Like;
 import com.hotnerds.post.domain.like.Likes;
-import com.hotnerds.post.domain.comment.Comment;
-import com.hotnerds.post.domain.comment.Comments;
+import com.hotnerds.comment.domain.Comment;
+import com.hotnerds.comment.domain.Comments;
 import com.hotnerds.post.domain.tag.PostTags;
 import com.hotnerds.tag.domain.Tag;
 import com.hotnerds.user.domain.User;
