+++ src/test/java/com/hotnerds/post/application/CommentServiceTest.java
@@ -4,8 +4,8 @@ import com.hotnerds.comment.application.CommentService;
 import com.hotnerds.common.exception.BusinessException;
 import com.hotnerds.common.exception.ErrorCode;
 import com.hotnerds.post.domain.Post;
-import com.hotnerds.post.domain.comment.Comment;
-import com.hotnerds.post.domain.comment.Comments;
+import com.hotnerds.comment.domain.Comment;
+import com.hotnerds.comment.domain.Comments;
 import com.hotnerds.post.domain.dto.CommentCreateReqDto;
 import com.hotnerds.post.domain.dto.CommentDeleteReqDto;
 import com.hotnerds.post.domain.dto.CommentResponseDto;
